03_MergeRequirements/requirements.txt
requests
python-cowsay
